refactor: Seperate view and logic
